Extract components 'RegularPriceMenuGroup' and 'RegularPriceMultipleSelectionMenuItem'
