+++ packages/js/product-editor/src/blocks/product-fields/regular-price/edit.tsx
@@ -8,12 +8,10 @@ import { useInstanceId } from '@wordpress/compose';
 import { useEntityProp } from '@wordpress/core-data';
 import { createElement, useEffect, Fragment } from '@wordpress/element';
 import { sprintf, __ } from '@wordpress/i18n';
-import { recordEvent } from '@woocommerce/tracks';
 import {
 	BaseControl,
 	// @ts-expect-error `__experimentalInputControl` does exist.
 	__experimentalInputControl as InputControl,
-	MenuItem,
 } from '@wordpress/components';
 
 /**
@@ -25,50 +23,8 @@ import { useCurrencyInputProps } from '../../../hooks/use-currency-input-props';
 import { sanitizeHTML } from '../../../utils/sanitize-html';
 import type { ProductEditorBlockEditProps } from '../../../types';
 import type { SalePriceBlockAttributes } from './types';
-import { TRACKS_SOURCE } from '../../../constants';
-import { handlePrompt } from '../../../utils';
-import { SetListPriceMenuItem } from '../../../components/variations-table/set-list-price-menu-item';
-import { VariationQuickUpdateMenuItem } from '../../../components/variations-table';
-import { VariationQuickUpdateMenuGroup } from '../../../components/variations-table/variation-actions-menus/variation-quick-update-menu-group';
-
-function isPercentage( value: string ) {
-	return value.endsWith( '%' );
-}
-
-function parsePercentage( value: string ) {
-	const stringNumber = value.substring( 0, value.length - 1 );
-	if ( Number.isNaN( Number( stringNumber ) ) ) {
-		return undefined;
-	}
-	return Number( stringNumber );
-}
-
-export function addFixedOrPercentage(
-	value: string,
-	fixedOrPercentage: string,
-	increaseOrDecrease: 1 | -1 = 1
-) {
-	if ( isPercentage( fixedOrPercentage ) ) {
-		if ( Number.isNaN( Number( value ) ) ) {
-			return 0;
-		}
-		const percentage = parsePercentage( fixedOrPercentage );
-		if ( percentage === undefined ) {
-			return Number( value );
-		}
-		return (
-			Number( value ) +
-			Number( value ) * ( percentage / 100 ) * increaseOrDecrease
-		);
-	}
-	if ( Number.isNaN( Number( value ) ) ) {
-		if ( Number.isNaN( Number( fixedOrPercentage ) ) ) {
-			return undefined;
-		}
-		return Number( fixedOrPercentage );
-	}
-	return Number( value ) + Number( fixedOrPercentage ) * increaseOrDecrease;
-}
+import { RegularPriceMenuGroup } from './regular-price-menu-group';
+import { RegularPriceMultipleSelectionMenuItem } from './regular-price-multiple-selection-menu-item';
 
 export function Edit( {
 	attributes,
@@ -172,150 +128,8 @@ export function Edit( {
 					/>
 				</BaseControl>
 			</div>
-			<VariationQuickUpdateMenuGroup
-				name="list-price"
-				label={ __( 'List price', 'woocommerce' ) }
-			>
-				{ ( { selection, onChange, onClose } ) => {
-					const ids = selection.map( ( { id } ) => id );
-					return (
-						<>
-							<SetListPriceMenuItem
-								selection={ selection }
-								onChange={ onChange }
-								onClose={ onClose }
-							/>
-							<MenuItem
-								onClick={ () => {
-									recordEvent(
-										'product_variations_menu_pricing_select',
-										{
-											source: TRACKS_SOURCE,
-											action: 'list_price_increase',
-											variation_id: ids,
-										}
-									);
-									handlePrompt( {
-										message: __(
-											'Enter a value (fixed or %)',
-											'woocommerce'
-										),
-										onOk( value ) {
-											recordEvent(
-												'product_variations_menu_pricing_update',
-												{
-													source: TRACKS_SOURCE,
-													action: 'list_price_increase',
-													variation_id: ids,
-												}
-											);
-											onChange(
-												selection.map(
-													( {
-														id,
-														regular_price,
-													} ) => ( {
-														id,
-														regular_price:
-															addFixedOrPercentage(
-																regular_price,
-																value
-															)?.toFixed( 2 ),
-													} )
-												)
-											);
-										},
-									} );
-									onClose();
-								} }
-							>
-								{ __( 'Increase list price', 'woocommerce' ) }
-							</MenuItem>
-							<MenuItem
-								onClick={ () => {
-									recordEvent(
-										'product_variations_menu_pricing_select',
-										{
-											source: TRACKS_SOURCE,
-											action: 'list_price_decrease',
-											variation_id: ids,
-										}
-									);
-									handlePrompt( {
-										message: __(
-											'Enter a value (fixed or %)',
-											'woocommerce'
-										),
-										onOk( value ) {
-											recordEvent(
-												'product_variations_menu_pricing_update',
-												{
-													source: TRACKS_SOURCE,
-													action: 'list_price_increase',
-													variation_id: ids,
-												}
-											);
-											onChange(
-												selection.map(
-													( {
-														id,
-														regular_price,
-													} ) => ( {
-														id,
-														regular_price:
-															addFixedOrPercentage(
-																regular_price,
-																value,
-																-1
-															)?.toFixed( 2 ),
-													} )
-												)
-											);
-										},
-									} );
-									onClose();
-								} }
-							>
-								{ __( 'Decrease list price', 'woocommerce' ) }
-							</MenuItem>
-						</>
-					);
-				} }
-			</VariationQuickUpdateMenuGroup>
-			<VariationQuickUpdateMenuItem
-				group="multiple-selections"
-				supportsMultipleSelection
-				onClick={ ( { selection, onChange, onClose } ) => {
-					const ids = selection.map( ( { id } ) => id );
-
-					recordEvent( 'product_variations_menu_pricing_select', {
-						source: TRACKS_SOURCE,
-						action: 'list_price_set',
-						variation_id: ids,
-					} );
-					handlePrompt( {
-						onOk( value ) {
-							recordEvent(
-								'product_variations_menu_pricing_update',
-								{
-									source: TRACKS_SOURCE,
-									action: 'list_price_set',
-									variation_id: ids,
-								}
-							);
-							onChange(
-								selection.map( ( { id } ) => ( {
-									id,
-									regular_price: value,
-								} ) )
-							);
-						},
-					} );
-					onClose();
-				} }
-			>
-				{ __( 'Set list price', 'woocommerce' ) }
-			</VariationQuickUpdateMenuItem>
+			<RegularPriceMenuGroup />
+			<RegularPriceMultipleSelectionMenuItem />
 		</>
 	);
 }
